Fix build, move LintItem defn, change literal to constant `kCodeViewGetDiagnostics`
