+++ apps/lsp/src/service/providers/diagnostics-yaml.ts
@@ -24,14 +24,10 @@ import {
 import { Document } from "quarto-core";
 
 import {
-  kEndColumn,
-  kEndRow,
-  kStartColumn,
-  kStartRow,
-  LintItem,
   Quarto,
   docEditorContext
 } from "../quarto";
+import { kEndColumn, kEndRow, kStartColumn, kStartRow, LintItem } from "editor-types";
 
 export async function provideYamlDiagnostics(
   quarto: Quarto,
